# [Core] Improve Tensor serialisation