This improves v1 tensor serialisation by directly relying on `torch.frombuffer` which removes the need for an temporary numpy array which is a bit easier to read.

Here's a small micro benchmark to verify that this is also faster:
```python
import numpy as np
import torch

from vllm.v1.serial_utils import MsgpackDecoder, MsgpackEncoder

encoder = MsgpackEncoder()
tensor_decoder = MsgpackDecoder(torch.Tensor)
numpy_decoder = MsgpackDecoder(np.ndarray)

array = np.random.rand(1, 3, 896, 896).astype(np.float32)
tensor = torch.tensor(array)
encoded_tensor = encoder.encode(tensor)
encoded_array = encoder.encode(array)
```

**main:**
```
Tensor encode/decode
5.44 μs ± 12.3 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
5.22 μs ± 9.13 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
Array encode/decode
5.46 μs ± 4.99 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
1.43 μs ± 4.35 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```

**This PR:**
```
Tensor encode/decode
5.03 μs ± 3.81 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
4.41 μs ± 7.52 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
Array encode/decode
5.02 μs ± 23.8 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
1.45 μs ± 3.2 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```